+++ assets/src/markdoc/components/Figure.tsx
