+++ assets/src/markdoc/components/Image.tsx
