+++ assets/src/markdoc/nodes/hr.markdoc.ts
@@ -1,4 +1,4 @@
-import HorizontalRule from '../../components/md/HorizontalRule'
+import HorizontalRule from '../components/HorizontalRule'
 
 export const hr = {
   render: HorizontalRule,
