+++ assets/src/markdoc/nodes/link.markdoc.ts
@@ -1,6 +1,6 @@
 // import { link as baseLink } from '@markdoc/next.js/tags'
 
-import Link from '../../components/md/Link'
+import Link from '../components/Link'
 
 export const link = {
   render: Link,
