+++ assets/src/markdoc/nodes/table.markdoc.ts
@@ -1,6 +1,6 @@
 import { Tag } from '@markdoc/markdoc'
 
-import Table from '../../components/md/Table'
+import Table from '../components/Table'
 
 export const table = {
   render: Table,
