+++ assets/src/markdoc/tags/callout.markdoc.ts
@@ -1,6 +1,6 @@
 import { Tag } from '@markdoc/markdoc'
 
-import Callout from '../../components/md/Callout'
+import Callout from '../components/Callout'
 
 export const callout = {
   render: Callout,
